+++ src/core/engineHelper.ts
@@ -10,11 +10,12 @@ namespace Lightning {
 
         protected _dpr:number;
         protected _renderer: PIXI.WebGLRenderer | PIXI.CanvasRenderer;
-        protected _world: PIXI.Container;
+        protected _world: Lightning.Group;
         protected _hud:HUD;
         protected _ticker:PIXI.ticker.Ticker;
         // protected _tweens:TweenManager
         protected _stateManager:StateManager;
+        protected _device:Device;
         protected _physicsManager:PhysicsManager
         protected _eventEmitter:EventEmitter;
         protected _storageManager:StorageManager;
@@ -80,7 +81,7 @@ namespace Lightning {
             this._renderer.backgroundColor = val;
         }
 
-        public get world():PIXI.Container {
+        public get world():Lightning.Group {
             return this._world;
         }
 
@@ -162,6 +163,99 @@ namespace Lightning {
 
         public get debug():Debug {
             return this._debug;
+        }
+
+        /**
+         * @description recursive pattern to loop over every child and recursivly loop over all of it's children and returning a count of them all from the root object.
+         * You must specify an object root
+         * 
+         * Example: 
+         * this.game.debug.displayCount(this.game.world);
+         * this.game.debug.displayCount(myContainer);
+         * 
+         * @see {Lightning.Engine}
+         * 
+         * @param rootObject 
+         * @returns {number}
+         */
+        public displayCount(rootObject:Lightning.DisplayObject):number {
+            return ((d) => {
+                var c = 0;
+                var r = function(d:Lightning.DisplayObject) {
+                    c++;
+                    for(let i of d['children']) {
+                        r(i);
+                    }
+                }
+                r(d);
+                return c;
+            })(rootObject);
+        }
+
+        /**
+         * @description recursive pattern to loop over every child and recursivly loop over all of it's children and returning an array of display objects that conform to any references.
+         * You may pass multiple references into a spread operator
+         * You can specify the starting root object
+         * 
+         * Example: 
+         * this.game.getObjectsByReference(this.game.world, 'alive', 'jimmy', 'fubar');
+         * this.game.getObjectsByReference(myContainer, 'onlyOneReference');
+         * 
+         * @see {Lightning.Engine}
+         * 
+         * @param rootObject 
+         * @returns {number}
+         */
+        public getObjectsByReference(rootObject:Lightning.DisplayObject, ... refs):Array<Lightning.DisplayObject> {
+            return ((d) => {
+                let result:Array<Lightning.DisplayObject> = [];
+                var r = function(d:Lightning.DisplayObject) {
+                    let ref:string = d.globalRef;
+                    for(var i of refs) {
+                        if(ref === i) {
+                            result.push(d);
+                        }
+                    }
+                    for(var t of d['children']) {
+                        r(t);
+                    }
+                }
+                r(d);
+                return result;
+            })(rootObject);
+        }
+
+        public sizeof(object:any):number {
+            var objectList = [];
+            var stack = [object];
+            var bytes = 0;
+
+            while ( stack.length ) {
+                var value = stack.pop();
+
+                if ( typeof value === 'boolean' ) {
+                    bytes += 4;
+                }
+                else if ( typeof value === 'string' ) {
+                    bytes += value.length * 2;
+                }
+                else if ( typeof value === 'number' ) {
+                    bytes += 8;
+                }
+                else if
+                (
+                    typeof value === 'object'
+                    && objectList.indexOf( value ) === -1
+                )
+                {
+                    objectList.push( value );
+
+                    for( var i in value ) {
+                        stack.push( value[ i ] );
+                    }
+                }
+            }
+            return bytes;
         }
     }
 }