+++ src/core/storage/storageManager.ts
@@ -4,25 +4,17 @@ namespace Lightning {
 
     export class StorageManager {
 
-        private _isLS:boolean = false;
+        private _isLS:boolean;
         private _map:{[key:string]:any};
-
-        public setItem:Function = this.setItemFallback;
-        public getItem:Function = this.getItemFallback;
-        public removeItem:Function = this.removeItemFallback;
-        public removeAll:Function;
-        public exists:Function = this.existsFallback;
-        public length:Function = this.lengthFallback;
+        private _forceNoLocalStorage:boolean;
     
-        constructor() {
+        constructor(forceNoLocalStorage:boolean = false) {
+            this._isLS = false;
             this._map = {};
+            this._forceNoLocalStorage = false;
+
             if(this.localStorageAvailable()) {
                 this._isLS = true;
-                this.setItem = this.setItemLS;
-                this.getItem = this.getItemLS;
-                this.removeItem = this.removeItemLS;
-                this.exists = this.existsLS;
-                this.length = this.lengthLS;
             }
         }
 
@@ -31,67 +23,12 @@ namespace Lightning {
          * @param key 
          * @param val 
          */
-        private setItemLS(key:string, val:any):boolean {
-            localStorage.setItem(key, val);
-            return true;
-        }
-
-        /**
-         * 
-         * @param key 
-         * @param val 
-         */
-        private setItemFallback(key:string, val:any):boolean {
-            this._map[key] = val;
-            return true;
-        }
-
-        /**
-         * 
-         * @param key 
-         */
-        private getItemLS(key:string):any {
-            return localStorage.getItem(key) || null;
-        }
-
-        /**
-         * 
-         * @param key 
-         */
-        private getItemFallback(key:string):any {
-            return this._map[key].val
-        }
-
-        /**
-         * 
-         * @param key 
-         */
-        private removeItemLS(key:string):boolean {
-            localStorage.removeItem(key);
-            return true;
-        }
-
-        /**
-         * 
-         * @param key 
-         */
-        private removeItemFallback(key:string):boolean {
-            if(this.exists(key)) {
-                this._map[key] = null;
+        public setItem(key:string, val:any):boolean {
+            if(this._isLS) {
+                localStorage.setItem(key, val);
                 return true;
             } else {
-                return false;
-            }
-        }
-
-        /**
-         * 
-         * @param key 
-         */
-        private existsLS(key:string):boolean {
-            if (localStorage.getItem(key) === null) {
-                return false;
-            } else {
+                this._map[key] = val;
                 return true;
             }
         }
@@ -100,33 +37,80 @@ namespace Lightning {
          * 
          * @param key 
          */
-        private existsFallback(key:string):boolean {
-            if(this._map[key]) {
-                return true;
+        public getItem(key:string):any {
+            if(this._isLS) {
+                return localStorage.getItem(key) || null;
             } else {
-                return false;
+                return this._map[key].val
+            }
+        }
+
+        /**
+         * 
+         * @param key 
+         */
+        public removeItem(key:string):any {
+            if(this._isLS) {
+                let item = this.getItem(key);
+                localStorage.removeItem(key);
+                return item;
+            } else {
+                if(this.exists(key)) {
+                    let item = this._map[key];
+                    this._map[key] = null;
+                    return item;
+                } else {
+                    return false;
+                }
+            }
+            
+        }
+
+        /**
+         * 
+         * @param key 
+         */
+        public exists(key:string):boolean {
+            if(this._isLS) {
+                if (localStorage.getItem(key) === null) {
+                    return false;
+                } else {
+                    return true;
+                }
+            } else {
+                if(this._map[key]) {
+                    return true;
+                } else {
+                    return false;
+                }
             }
         }
 
         /**
          * 
          */
-        private lengthLS():number {
-            var _lsTotal=0,_xLen,_x;for(_x in localStorage){_xLen= ((localStorage[_x].length + _x.length)* 2);_lsTotal+=_xLen; console.log(_x.substr(0,50)+" = "+ (_xLen/1024).toFixed(2)+" KB")};console.log("Total = " + (_lsTotal / 1024).toFixed(2) + " KB");
-            return _lsTotal;
+        public removeAll() {
+            if(this._isLS) {
+                localStorage.clear();
+            } else {
+                this._map = {};
+            }
+        }
+
+        public length():number {
+            if(this._isLS) {
+                localStorage.length
+            } else {
+                return Object.keys(this._map).length;
+            }
         }
 
         /**
+         * @description Detects if local storage is avaialble in the browser
          * 
+         * @returns {boolean}
          */
-        private lengthFallback():number {
-            return Object.keys(this._map).length;
-        }
-
-        /**
-         * 
-         */
-        private localStorageAvailable() {
+        public localStorageAvailable():boolean {
             var a = 'a';
             try {
                 localStorage.setItem(a, a);