+++ src/core/timer/timer.ts
@@ -66,14 +66,15 @@ namespace Lightning {
         }
 
         public destroy() {
+            this.game.ticker.remove(this.update, this);
             this._active = null;
             this._autoDestroy = null;
-            this.game.ticker.remove(this.update, this);
             this._currentTime = null;
             this._events.destroy();
             this._interval = null;
             this._isLoop = null;
             this._lastTick = null;
+
         }
 
         public get events():EventEmitter {